Moved StepFunctions into own namespace and file
